src/face_recognition/requirements.txt
requests>=2.28.0
opencv-python>=4.8.0
python-dotenv>=1.0.0